mcp-sqlServer/tests/test_requirements.txt
# These should already be in your main requirements.txt:
# pyodbc
# azure-identity
pyodbc>=5.2.0
azure-identity>=1.12.0
fastapi>=0.100.0
pydantic>=1.10.10
uvicorn>=0.23.0
azure-kusto-data